Added support for theme handling with sass css compilation Refactored Portlet plugin to reuse the sass compilation Added unit and integration test for theme handling
