+++ src/integTest/projects/servicebuilder/build.gradle
@@ -8,9 +8,7 @@ repositories {
 
 buildscript {
     dependencies {
-        //classpath files("${buildDir}/../../../libs/liferay-plugin-0.1.0-SNAPSHOT.jar")
-
-        classpath files("${buildDir}/../../../../../build/libs/liferay-plugin-0.1.0-SNAPSHOT.jar")
+        classpath files("${buildDir}/../../../libs/liferay-plugin-0.1.0-SNAPSHOT.jar")
     }
 }
