+++ src/integTest/projects/theme/src/main/webapp/css/main.css
@@ -1,0 +1,4 @@
+$color1: #fff;
+
+body { color: $color1; }
+
